Ban helper OOP pattern
We see a lot of "helper" OOP objects being introduced but have found
that these are often very difficult to maintain. Their usage can spread
out to many different files and it's often unclear what the goal of the
helper class is beside deduplicating some code. This makes refactoring
it without breaking consuming code very difficult.

This commit extends our best practices workflow to flag helper classes
being added (we ignore existing ones being modified) and also extends
our workflow with a GitHub script that will post a message explaining
both the config overrides and the helper method errors. The script
creates a comment that it keeps updated based on the status and will
clean up the comment when the issues are resolved.

The script is introduced as a standalone JavaScript file which allows it
to be more easily tested during local development, this prevents a lot
of the needed back-and forth pushing on our CI.