Migrate lazy loading of iFrames to Embed (where we use iFrames)
We used to do this using the lazy contrib module, but the only thing that it does these days
is use native browser attribute "loading=lazy".
Considering we're mainly using iframes in Embed, and images already migrated from social_lazy_loading to the Drupal Core native version
we don't need a social_lazy_loading module anymore.

So we create our own version of the lazy text processor, that adds the loading lazy attribute to all iframes.
This has a weight of 999, so it runs at the end, which is helpful because when we embed content it isn't an iframe until it passes
that processor.