Add PR title validation to PR manager
This was the only validation that Mergeable still had that the PR
manager was missing.

We've also changed it slightly so that we get rid of `Issue #PROD` which
is now very prolific but which is pushing Drupal standards onto our Jira
issues. We provide a specific message for that case to guide people to
the new format.

The benefit of making sure we properly distinguish is that other tools
looking at titles know whether to check Jira or Drupal.org for the
source of truth.