Add Configuration to influence the order of the bottom bar features
This gives admins a way to deal with the bottom bar features for the App.
This is just a proof of concept, later on we can add SM features when those details are clear.
Right now we use the existing hook as input for the ConfigForm which preferred features are available.
We add a draggable table to the form, which shows all available features.
And we change the place where the hook is invoked, for GraphQl, to the configuration from the form submitted.