Implement cacheability propagation for secret stream wrapper
Drupal does not provide a way to propagate cacheability information from
a stream wrapper to the caller and above. I came up with a workaround
using the renderer to propagate the max age of the link to a surrounding
render context so that the dynamic context can be updated.

We also expose it in the Drupal settings attachment so that we can
propagate it to a response subscriber which sets the caching for the
entire response.