Move the secret file system to a separate module
This increases maintainability because it makes everything around the
secret filesystem into a standalone module that doesn't need anything
else to function. We do introduce it as a key dependency of social_core
so that it can be used by all of Open Social's modules. Downstream sites
can opt-out by changing the field config for individual file fields.