Make Text Format filter aware of Secret Files
Drupal contains a filter that limits images in formatted text to the
platform domain. This filter has special handling for private files and
treats all other wrappers as publicly accessible. However, this causes
the filter to reject images from the secret file system.

We overwrite the plugin to inject our own logic which is slightly hacky
but also the easiest way to get this working. Unfortunately we can't
compose the filter because by the time Drupal's filter returns, the
damage is already done.