Replace DrupalContext usage with own implementation in Behat
DrupalContext provides quite a few ways to create content but doesn't
actually validate the input which can lead to unexpected test failures.
One place where it's indeterminate is with authors.

For most things we already have ways to create scaffolding of entities
with much better feedback. This commit replaces the content parts of
DrupalContext that we were still using and some older
SocialDrupalContext methods with the more modern implementation.

This is a first step towards being able to remove the DrupalExtension.