Issue #3454972: Enrolment & Invite pages in Manage Members in Events and Groups does show selected users
On an event's or group's "Manage members" tabs, when we select users in an event, it does not show how many users are selected. Also, count remained 0 on selections as well as on page refresh.

This was happening because we were calculating the total count is indifferent logic from VBO's own parent class. So, we are aboting the same solution across all the VBOForms used in Open Social.