IOSAP-457: Provide external identifier field
Also PROD-29703.

As a developer I want a re-usable externalId implementation,
so that I can easily apply this to different entity types.

This is achieved by creating a new field type with 3 subfields:
- External ID (external_id, string)
- External Owner Target Type (external_owner_target_type, string)
- External Owner ID (external_owner_id, integer)

The field storage settings can be used to determine which entity types
can be referenced. The field provides a getExternalOwnerEntity()
method that returns an entity object based on the based on the external
owner target type and external owner ID.

Constraint classes have been determined to validate input based on the
business logic required for the field, and therefore it is important to always
validate entity/field before it is saved programmatically.

Decision records for why the dynamic entity reference or
core entity reference field types were not used, are in PROD-29703.