IOSAP-457: Add external identifier field for user
Add an external identifier field to the user entity.
This field is locked, as it should not be managed
via the UI, but rather by modifying the configuration.
The field is provided without any settings and is,
therefore, unusable by default. At least one target
type needs to be specified by modifying the field
storage configuration to ensure the field can be
properly used.

Include hook update to ensure that field is created
also for existing installations.

Field usage example:
SCIM provides the concept of externalId which
can be used to identify a resource with an identifier
that is known to the SCIM client. This prevents the
SCIM client from having to store a mapping between
its internal data and our ID, instead putting that burden
on us as a service provider.