Issue #3458854: Fix url in getRelatedEntityUrl if entity doesn't have canonical link template
Sometimes the entity related to the activity does not have a canonical URL,
which causes a fatal error when viewing the activity.

This error occurs because there is no check to see if the entity has a canonical LinkTemplate.

So, we need to add  a condition to check if entity hasLinkTemplate canonical