Replace grequest with contrib version 2.2.2
This commit replaces the custom module, which came from a patch that was
introduced before Group 1.4, with the current 2.2.2 version of the
grequest module. The next steps are to then remove the module and pull
it in through composer, but this commit gives a clear overview of the
differences between our custom version and the contrib version at the
time of the switch.