Move to contrib drupal/grequest module
This replaces the in-repo copy of grequest with a copy that's installed
with Composer from the Drupal packagist. This ensures that dependencies
are also properly pulled in.

We choose 2.2.2 which is the latest stable version and we lock it to
match the current policy of other dependencies in the distribution.