Add test coverage for flexible group request to join
This commits adds test to document the behaviour of the "Request to
Join" feature for flexible groups after discussing the behaviour with
our PO and Designer.

The join functionality is moved to a new folder under "groups" in the
testing structure. The outsider view test was previously modified to
have a bit of test code around request to join and direct joining. This
was moved to their own test so that the feature for viewing is really
only for viewing and the joining features are separate.

The join functionality is not placed under flexible because after the
group migration we're only left with a single group type. The
functionality for other group-module-based things are different
features.