Move Flexible Group membership request config to flexible group module
The config for the membership request functionality for flexible groups
was included in `social_group_request`. However, the
`social_group_request` module is a module whose purpose is to provide
building blocks on top of `grequest` which other Open Social group types
can re-use (e.g. discussions or courses).

By moving the functionality into `social_group_flexible_group` we bring
that module closer to containing all the (optional) functionality that
is specific to the Flexible Groups _feature_, while keeping only the
generic shared behaviour in social_group_request for the membership
request _system_.