Return user as offline when the session table isn't exist and app is executing by cli
In the Drupal 10.3 version the session table isn't be created by system_schema, so the table is created when a session is created.
To execute the PHPUnit and install the Open Social using shell-script the cron is executed and the activity job check if the user is offline to send the notification and as it is a new installation the code broken because the session table is missing.
More details about session table: https://www.drupal.org/node/3431286