Allow custom group types to be configured.
We used to load all roles / group types based on if -outsider is part of the name.
This worked fine for a lot of the circumstances of the old group type where this was the case.
Now, with the move to Group v2, oustider became a scope of the actual group role.
So to make this compatible, we load all group roles, and figure out if we have a role with outsider scope that has
the permission to join groups. Any group types without an outsider scope and or the join group permission we can't configure.
As that means there is no group role specified that is able to quickjoin.