Add compass file to all top-level modules
The compass file is used in Compass to describe components that make up
our software. We initially define two types of packages: libraries which
describe base systems that other things can build on top of; or
capabilities which indicate something is a feature that users will be
able to see and interact with.