Issue #3481407: Cannot save alternative frontpage changes
In https://www.drupal.org/project/social/issues/3468174 we added a validation handler. This validation handler is also triggering on submission of form at /admin/config/alternative_frontpage which is causing an issue on `$path_alias = end($path_alias);`

So, we have added a check in form alter that the validation should only trigger on content entity's which are nodes and groups in our case.