Change compass.yaml contents to match Compass tool
The contents of the compass.yaml files that were manually added differed
from what the Compass tool would generate through the UI. This could be
the reason that the components are not being imported (the tool doesn't
provide proper feedback about this).

This PR changes the order of fields to be what comes out of the
"Config-as-code" tab in the tool so that copy pasting from the tool to a
file yields consistent results. It also changes `tier` from `null` to
the lowest tier which is `4` since it's not clear whether `tier` may be
empty (you can't make it empty through the UI). Finally it changes
`customFields` to `null` from `[]` which differs from the documentation
but matches what's automatically dumped.

While we're add it we add two links: one to the repository this
component lives in (the distribution) and the other to the Jira project
that work lives in for our component (Product).