Recommit default-calendar.svg icon
In the distribution the `.gitattributes` file that was provided by
Drupal would be installed outside of the git repository that contained
the distribution itself (`root/.gitattributes` with the git directory at
`root/html/profiles/contrib/social/.git`). This meant that it would not
be applied.

With the move into Cable Car the `.gitattributes` file is now present
for the entire CMS application which includes the install profile. This
causes git to apply the attributes to a file in one of our modules.