PROD-31686 Merge "temporary" translataibility patch into code
For now we commit the code with a @todo that points to removing the
ConfigOverride (which solves the underlying issue) and this can be
done in https://getopensocial.atlassian.net/browse/PROD-22612.

Unfortunately because of the large number of patches I don’t have
the time to do the config override removal now.