Merge install profile PHPStan rules into Cable Car
This ensures that the install profile is now also covered by PHPStan. It
enables a rule for private function calls that was added to the
distribution but not yet enabled in Cable Car. Unfortunately there are a
lot of existing issues in Cable Car which are grandfathered into our
baseline.

To keep separate track of existing issues in Cable Car and in the
distribution we simply include the distribution's existing
phpstan-baseline.neon rather than merging the two. That baseline should
only get smaller from here on out.

We also update our open_social_dev to get the prophesize PHPStan
extension and make PHPStan understand our mocks. This fixes some uses of
prophesize which were unfortunately ignored.

The addition of the entity mapping also allows us to remove or update a
bunch of ignored PHPStan errors (which were previously a generic
EntityInterface ignore rule).