Merege PHPUnit configuration from install profile into Cable Car
This updates Cable Car's PHPUnit file to also run the PHPUnit tests in
the install profile. We start by making our PHPUnit config a bit
stricter (warning for output and empty tests) which was never enabled in
Cable Car but are code smells (and seem to not currently be a problem,
so lets keep it that way).

We keep the current "all" test-suite but rename it "saas" and add the
install profile's test suites specifically. This still allows running
them separately if someone so desires. We can rethink the structure in
the future.

Finally the coverage line is moved to the bottom because usually when
opening phpunit config you care about the php config and rules first.