Move CSpell configuration to Cable Car
Cable Car didn't have a CSpell configuration yet, it was only in our
install profile. This means we now move the configuration over so that
we have it accross our repository.

We move this to the root of the repository since it's actually a tool
that works across languages and proper spelling is important regardless
of our product.