DE-549 Hardcode group types to work around config overrides
In some cases config overrides are executed in a way by Drupal that the
social_group_types alter does not collect the right group types. We
hardcode the affected ones until we've gotten rid of the config
overrides and of the group type hook itself (which should be replaced by
group type configuration).