Issue PROD-31477: Fix bulk operations for admin/content and admin/content/comments
In our Distro we use the field bulk_form to perform actions on nodes, but when social_content_export gets enabled a new field is introduced, causing issues with dissapearing options. This commit fixes the issue by using the same field for the Distro so we only need to extend it with the social_content_export.
This includes:
- Change of the config install for the views.view.content
- Update hook in the social_core to update this view based on the options they had selected for the old field
- Change of the config install for the views.view.comment
- Update hook in the social_comment to update this view based on the optiosn they had selected for the old field
- Adjust the config modify to only extend the original view and not replace it
- Remove the override in the .module files