Allow Cable Car Behat config to run install profile tests
We want to be able to easily run all of our Behat tests including the
one from the install profile that are not part of our default config.

To achieve this we include the install profile Behat files into the main
config in Cable Car and remove the config from the install profile.