PROD-30735: Replace the logic for generating the add form page titles for entity types Node, Group, Media, Post and Organization
The old logic would not always be grammatically correct, as the article was not tied to the next word in english. The current logic ties both so the article is dynamically set according to the next word.

The hooks `hook_social_core_title` and `hook_social_core_title_alter` are deprecated and will be removed in open_social:14.0.0.