Fix issue with social_event_addtocall images not being added to emails.
When social_event_addtocal is enabled it will create a snippet to include in emails. For security reasons our mail_html format restricts outside images. This has a check which does not work for SVG and therefore always blocks the images. There is a patch available on https://www.drupal.org/project/drupal/issues/2998318 which would first needs a security review as SVG can be prone to XSS. When looking at our implementation svg is made for scaling which we don't need. So replace it with webp to fix the issue.