Ensure secret file test is resilient against slow test runs
The test creates a request which sets a cache lifetime for the request
time + 42 seconds. However, the response cache subscriber converts this
to a header which uses the current time. So if the request takes 2
seconds, then the browser will be told that that 2 seconds has elapsed
and the resource is cacheable for shorter.

This is the desired behaviour but it creates a race condition in the
test where the cache lifetime of the secret files is set on one second
and the cache subscriber happens the next second. That causes the
lifetime to be slightly lower than 42 (e.g. 41 or 40) which causes the
test to fail.

Because the behaviour is correct and the time changing between what's
provided and what is returned to account for the request is part of the
expected behaviour, we slightly adjust the assertions to allow a wider
range of values. If we have a result that's lower than 40 then the test
took longer than 2 seconds, which would indicate a serious problem in
our test. At that point it would be okay for the test to fail.