Migrate the SaasGroupSettings to our Distro SocialGroupSettings
This form already extends the current SocialGroupSettings Form.
No additional permissions or anything required, the only thing it does it allows to decide if VU can actually create Groups of a certain Group type and adds this to the Group Settings existing, with some CSS and JS love which for now we migrate with it, we can iterate in the future to improve it (including schema file updates and more).