PROD-30185: Added query condition to check published status for nodes
If a user follows a taxonomy term, it receives notifications for content created with that tag. If this tag is added to an unpublished content, the user still receives notification. This shouldn't happen and to address this we added checks in ActivityContext 'FollowTaxonomyActivityContext::haveAccessToNode()' to add published nodes only for notifications.