Move the SaaS themecontroller to the Distro to make sure our SM can only change the Open Social theme settings
In our distribution we used to have SM have the ability to change everything related to themes by the power of the administer themes permission.
in our product version, a SM is only able to change the theme settings for SocialBlue (which uses improved theme settings to allow changing colors and radius and such).
We want to make our distribution roles match the permissions of our product, so we revoke the administer permissions, create a new one, and only present that to SM.
Other than that the code is copied one on one, so there is minimal chance of regression and we can improve the controller later on.