Also move the preprocess page, so that SM can not switch to other themes.
This should preprocess the admin page for themes, we want the SM to only see the socialblue settings there.
As we also remove the administer themes permission there, we switch the uid 1 check to actually check for that theme.
That should work for any existing Distro user, as that is the common permission to check (we dont ahve individual theme permissions)