Ensure all permissions for VU are granted matching social_saas_permissions, on install.
When we enable social_saas_permissions we reset all permissions.
Most of these permissions are already given by the modules that enable them, so if we don't reset the permissions we only have to care about removing permissions that are enabled too much, or aren't enabled on install at all.

In the PR connected you can find a `diff` showing the config export of the permissions of VU users based on the install of social_saas_permissions and social_saas_core, and based on social_saas_core without social_saas_permissions, that should tell us which permissions we need to grant or revoke.
The big diff is there, because in the product we revoked a lot of permissions for AU and moved them to Verified, but we did not do that for saas_permissions :( luckily this fixes that for new installs.

We only do this for install to minimize regression, but create value for new installs.
Because we still have extensions that allow changing permissions by SM, resetting those on update do not make sense as we will have to do that later on again and again until we stop that from happening.