We've removed the administer block permissions, as it was gone in saas permissions, causing issues with landing pages and custom content list blocks.
Instead of putting back administer blocks, which is like a catch all permission, we found that adding permissions just for:
- the AN hero block (which is the one on the default AN homepage) for editting
- the block library (which is the content blocks overview that just lists all the custom blocks created)
- the custom content list CRUD

Allows us to hand out permissions so SM can edit the AN hero block
and CM and SM can create custom content lists so it can be referenced on a landing page.

For dashboards there are different permissions, that doesn't need the block layout or block library at all or individual block_content type permissions for that matter.