Include default permissions that we used to set manually in our Setup Demo / Tier 2 platform document.
Because of our former `social_saas_permissions` and it revoking all permissions and granting all of them itself (see PROD-1111)
it means we had to do a lot of manual drush cset for permissions that were incorrectly revoked.
Now we finally are able to move everything to the correct extensions and when they are being installed, so we can also move all the missing permissions from the setup document
and ensure they are dealt with when enabling the module. These permissions were missing on install of the module.