Assign permissions removed by social_saas_permission routesubscriber
in social_saas_permission we were setting requirements for different routes.
Making sure instead of handing out those permissions, we create our own weird dashboard permission.
That also meant that the right permissions we're removed for SM and should've been added on update for some of those.

So in this commit I'm re-assigning those permissions.
But also create a RouteSubscriber for those paths (like site information / menu links) that have permissions that give away to much.
For example administer site configuration is also used on cron and other system pages.