Make sure as VU you can't view an unpublished group, as SM or Group Manager you can if it's your own group.
Background info:
We noticed this issue didn't happen when social_course enabled; we seemingly have done our own implementation of access handling for published/unpublished there but actually we were changing all access checks for all the selected group routes (regardless if it's flexible, course or what not).

This fix originated at the start of courses 7 years ago (before it landed in Groups):
https://git.drupalcode.org/project/social_course/-/blob/5.1.x/social_course.module?ref_type=heads#L1896
https://git.drupalcode.org/project/social_course/-/blob/5.1.x/src/Controller/GroupController.php?ref_type=heads
https://git.drupalcode.org/project/social_course/-/blob/5.1.x/src/Routing/RouteSubscriber.php?ref_type=heads#L56

And we never made the product implementation of Group compatible with the access checks for published/unpublished: https://www.drupal.org/project/group/issues/2873212
As the permissions we set on the overviews are things like “view published group” and handed out to any outsider it means the actual page is visible, access is not determined whether or not the actual group is published or unpublished or not for those views page. The actual Group as an entity related details are correctly hidden when its unpublished, that is why the description doesn’t show up for the mentioned bug. The page only shows blocks, that obviously also don’t take the published status in to account. (edited)

We also have other access checks for flexible groups, like the FlexibleGroupAccesCheck (for group visibility) and a hook_entity_access for groups as well.
So for now I chose to make this similar to courses and we get to fix it for the entire product and all group objects better together.