PROD-30018: Create upgrade path to contrib grequest module
There's a few differences between what was in our grequest module and
what is in the version that's now in contrib that we're moving to. This
requires some config and databae updates.

These updates are placed in `social_group_request` under the assumption
that all platforms that use `grequest` do so through this Open Social
integration module.

The upgrade path puts an updated Open Social as close as possible to a
fresh install.