PROD-30741: Added AccessControl group relationship handler for social_collaboration
From Group v1 and above, the module restructured how access check works. This is elaboratively explained at https://www.drupal.org/docs/extending-drupal/contributed-modules/contributed-module-documentation/group/group-v1-guides/extending-group-v1s-access-control-model#s-an-example-and-groups-default-access-control.

Because Group returns a hard AccessResult::forbidden() result for grouped content, it is not possible for a custom module to simply implement hook_node_access() or hook_entity_access() and return a AccessResult::allowed() result as the AccessResult::forbidden() will win.

So, we require our own AccessControl relationship handler for the entity access which are posted within group so that co-authors can edit a node.