ISAGOAM-492: Introduce user affiliations
User affiliations is a feature, that allows members to determine
which organization they represent in the community. Users can
define this through their the profile settings.

Affiliations can be enabled on a per-platform basis. When activated,
the "Function and organization" section in profile settings is
replaced by "Affiliation representation."

The affiliation section in profile settings includes two parts:
Group affiliations and non-platform affiliations. Platform affiliations
are always available, while group affiliations are displayed only
if there is at least one eligible group affiliation. Affiliation candidates
refer to group types that are eligible for affiliation, and they can be
enabled per platform.

The rules for displaying group affiliation options in user profile
settings are as follows:
- The group type must be an affiliation candidate.
- The group type must have affiliation enabled.
- The user must be a member of the group.
- The group must be published.
- The group’s visibility must not be restricted to members (for flexible groups only).