PROD-33190 Fix memory leak in Behat test runner
The AvoidCleanUpTrait was used to ensure we don't run database queries
at the end of each scenario, since we start from a fresh database for
each test so that's just wasted time cleaning up a database that gets
discarded.

However, while disabling that work, the cleanup of any arrays that point
to PHP Objects were also disabled. This means that in test runs with a
large number of scenarios these arrays may grow very large since
contexts stay around between scenarios. With our new Docker set-up we're
limiting any CLI processes to 1 GB in the CI which is now showing this
leak.

The AvoidCleanUpTrait is adjusted so that we clear out the arrays. This
should ensure the references to the objects are removed and let PHP
garbage collect them to reduce the amount of memory the test runner
uses.