PROD-29627 Introduce Drush test bridge for module management
In our current Behat set-up we're calling into Drupal directly. However,
this has quite a few downsides:

1. The tests are coupled to a specific version of our application which
means we can't write automated upgrade path tests that upgrade in the
middle of a test.

2. Drupal expects global state to be short-lived within a request.
However, Behat runs multiple tests within the same state which can cause
issues with Drupal.

This commit creates a Drush command that can be the long-runned process
and can be easily restarted. The command allows communication through
stdin/stdout in JSON format. This provides a JSON based API that can be
stable between versions which solves the upgrade path issue. The ability
to easily restart allows clearing the process' caches during things like
module installs.

We don't add any functionality into it yet and do so in follow-up PRs.
However, we do change modules being installed from happening in the
Behat test runner, to being done in Drush. This should fix a memory
leak issue that was experienced in some tests because the install now
happens in a separate process that's cleaned up at the end.