Apply config link patch for GraphQL module
This pulls in the bugfix for https://www.drupal.org/project/graphql/issues/3506885
which is released in a newer version (thus the commit we link to is
stable). However, we're blocked from upgrading.

While installing the GraphQL module with Drush 13 the task fails with
the error
> Route "graphql.config_page" does not exist.

Adding this patch ensures the route is fixed and the error doesn't
occur.