Ensure comment is posted by the right user in test
Before the memory leak change was applied the current user was not
properly reset between tests. This caused the `current user` in
post-comments to apply to the `PostUser2` which was still logged in from
the previous scenario.

After fixing the test isolation with the memory leak, there's no longer
a user logged in which causes an error when the `current user` is
anonymous.

Reading the rest of the test confirms that the user (who is "verfiied")
is editing their own post, so it makes sense that they should be the
author and we make this explicit in the `Given comments` step.