Make use of Mailpit API in Behat tests
Previously, tests would require the `@email-spool` annotation in order
to be able to test emails. This annotation would then modify
configuration in our application to change how email sending works.

It's considered bad practice to have tests change configuration of an
application in a way that is not representative of production (as is the
case here). Additionally, while trying to refactor how our test runner
worked we ran into issues with race conditions in how this configuration
was managed.

In a previous commit we've introduced Mailpit as the service that our
application sends email to during development and testing.

In this commit the testing context is changed so that it no longer
modifies how our application works. It now uses the Mailpit API to query
the Mailpit server and figure out if there are emails that are relevant
to the test. This has an added benefit that we no longer need to do
complex file management and decouple our testing from Symfony Mailer.

These same Behat steps can now also be used to test other applications
(like Engagement Studio) if those applications are properly configured
to send their emails to Mailpit.

The email cleanup from the test is removed on purpose. We implement a
timestamp based system to make sure we ignore previous emails in a test
and assume that in our CI all of the testing infrastructure is removed
after testing. This changes makes it easier locally to review emails
that are being sent as part of certain scenarios. For example, for a QA
test it's now possible to run a test scenario and review the styling of
the emails involved.