Run the queue processing for email twice
We found that for tests where a lot of emails are being sent, the single
process run might not be able to pull all the emails through the queue.

To solve this, for now we run the queue process function twice. In the
future we can refactor how these kinds of tests work to make a "Wait for
email" step or similar, which can run to completion or a timeout, rather
than an arbitrary chosen size.