Fix IATA request management test
The test would sometimes pass and sometimes fail. The functionality was
actually broken and the passing test would occur based on the test
clicking on the incorrect "Remove" button in the list. This would cause
the Margot Robbie relationship to be removed which had the the proper
form alters implemented.

To fix this, two changes are made:

1. We implement a "Should see" step which validates that we're actually
removing "Ryan Gossling" and not someone else.
2. We make a more robust test step to select buttons in rows of a table.

The new test step is marked as `ACCESSIBILITY ISSUE` because an
accessible button can be selected without ambiguity which is clearly not
the case here. Note that the inacceessibility is not in the test but in
the implementation (so it was also present with `xth 'n'`, which we
should also give a similar warning text in a separate PR).

Fixing the actual accessibility issue in the code is out of scope for
what we have time for but now at least the test step makes this clearer.