PROD-33412: Prefix the Project ID as source to each event payload
The EDA Handlers all set the path rather than the full URI as Source for the event. E.g. /user/login or /node/add/event.
That means we can safely add the source of including the unique project ID before the path and afterwards make the event immutable. So it'll become projectid.opensocial.site/user/login or projectid.opensocial.site/node/add/event.
The project ID from the setting.php file, `project_id` is used, but it defaults to an empty string.

The DispatcherInterface is updated to use the V1 CloudEventsInterface, as it contains the actual methods we're expecting rather than just the specVersion().