Use drush cr to clear redis cache on Behat database load
In previous commits we added the settings.php to our monorepository. It
turns out that in the previous set-up there was actually no connection
set up to Redis and the service wasn't used in local development or our
CI.

With the connection to Redis in our CI it was revealed that although we
clear caches in the Behat process we don't clear caches outside of our
database, so many tests were failing because Redis had cached data from
the previous run.

We introduce a `drush cr` call to ensure that all caches (database or
external) are cleared after loading a fresh test database. The separate
clear cache steps are kept because they ensure the memory cache in Behat
is cleared. This causes some duplicate work, but that can be removed
when we move data handling outside of Behat.