ICSO-188: Introduce user profile segment rule
Introduce user profile user segment rule (plugin).
Introduces the mechanism for defining dynamic segments
based on user profile information, such as the role.

The user segmenting feature that we're building allows users
to be selected based on their properties and relationships and
subsequently outcomes of various things to do with these selections
can be configured. A rule plugin is added that makes user profile
properties available for use in segment creation.