Set up the Context for UserSegments within Behat.
This is the groundwork done for the creation of UserSegments in Behat.
We will likely remove/alter the current implementation, but this serves as a showcase the steps actually work.
Once we have the UI working, and the full storage layers there, we're able to better write tests that cover working with segments.