Add test coverage for viewing content with user segments on a node and in the overviews (search will come later).
This ensures we can create our content using a segment visibility.
We use the ProfileUserSegmentRule, allowing us to use user roles to create segments.
We're using a new role, as CM and SM (our default roles) have bypass permissions, so using those roles doesn't make sense.

We added contexts to connect segments to content by label, although the labels aren't unique, and in the frontend we render them by ID, so that's why we need to tie them together on the EntityTrait (this is the easiest so it will work for any Entity that uses this segment field).
This is done for Dashboards, Discussions, Events and Topics to make sure we have a good selection of the distribution and our extensions.