Updates and adds the event payloads for event enrollment related events
The previous implementation added incorrect statuses for the different event enrollment related events. This is fixed here, and thus added at once in a single commit.
Two existing hooks for entity_insert and entity_delete are adjusted to be specific to the events they trigger. Others are added in the controllers.

This covers:
- PROD-33361
- PROD-33362
- PROD-33363
- PROD-33364
- PROD-33365
- PROD-33366
- PROD-33367
- PROD-33368
- PROD-33369